+++ www/gatsby-config.js
@@ -56,6 +56,7 @@ module.exports = {
       options: {
         plugins: [
           `gatsby-remark-graphviz`,
+          `gatsby-remark-code-titles`,
           {
             resolve: `gatsby-remark-images`,
             options: {
@@ -104,14 +105,13 @@ module.exports = {
     `gatsby-transformer-csv`,
     `gatsby-plugin-twitter`,
     `gatsby-plugin-react-helmet`,
+    `gatsby-plugin-sitemap`,
     {
-      resolve: `gatsby-plugin-google-tagmanager`,
+      resolve: `gatsby-plugin-google-analytics`,
       options: {
-        id: `GTM-KLZLVML`,
+        trackingId: `UA-93349937-5`,
       },
     },
-
-    `gatsby-plugin-sitemap`,
     {
       resolve: `gatsby-plugin-feed`,
       options: {
@@ -185,5 +185,16 @@ module.exports = {
     },
     `gatsby-transformer-screenshot`,
     `gatsby-plugin-subfont`,
+    // {
+    // resolve: `gatsby-plugin-guess-js`,
+    // options: {
+    // GAViewID: `142357465`,
+    // // The "period" for fetching analytic data.
+    // period: {
+    // startDate: new Date(`2018-1-1`),
+    // endDate: new Date(),
+    // },
+    // },
+    // },
   ],
 }