+++ packages/gatsby-transformer-yaml/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.1-rc.2"></a>
+
+## [2.1.1-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-yaml/compare/gatsby-transformer-yaml@2.1.1-rc.1...gatsby-transformer-yaml@2.1.1-rc.2) (2018-09-05)
+
+**Note:** Version bump only for package gatsby-transformer-yaml
+
 <a name="2.1.1-rc.1"></a>
 
 ## [2.1.1-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-yaml/compare/gatsby-transformer-yaml@2.1.1-rc.0...gatsby-transformer-yaml@2.1.1-rc.1) (2018-08-29)